Pull help "logic" out of config.go
